refactor: Move `extract` function to util submod
