+++ src-tauri/src/northstar/install.rs
@@ -5,7 +5,7 @@ use std::{cell::RefCell, time::Instant};
 use ts_rs::TS;
 
 use crate::constants::TITANFALL2_STEAM_ID;
-use crate::{extract, GameInstall, InstallType};
+use crate::{util::extract, GameInstall, InstallType};
 
 #[cfg(target_os = "windows")]
 use crate::platform_specific::windows;
